refactor: split spdx_visualize.py (1207 lines) into 7 modules (all <400 lines)
Extracted into app/viz/ package:
- extract.py: graph extraction from SPDX JSON (~234 lines)
- styles.py: CSS template (~140 lines)
- html_parts.py: header, legend, UI HTML fragments (~126 lines)
- js_simulation.py: D3 force sim, layout, tree/BFS, fanout (~336 lines)
- js_interaction.py: rendering, tooltips, drag, highlight, search (~272 lines)

Main spdx_visualize.py reduced to ~194 lines (CLI + orchestrator).
JS code no longer needs {{/}} f-string escaping — much more readable.
All 670 tests pass, HTML output functionally identical.
