+++ app/viz/__init__.py
@@ -1,0 +1,3 @@
+"""Visualization sub-package for SPDX dependency graphs."""
+
+from .extract import extract_graph  # noqa: F401
